+++ Sora/ConnectionController/ConnectionNavigationController.swift
@@ -1,8 +1,6 @@
 import UIKit
 
 class ConnectionNavigationController: UINavigationController {
-
-    weak var connectionController: ConnectionController!
     
     override func viewDidLoad() {
         super.viewDidLoad()